Fixed PPs not consuming while confused.
This was caused by earlier commit fixing confusion damage while
recharging. Unfortunately, turnMemory was checked for "NoChoice" with
the indexing operator[] which automatically creates a value there. Other
place where PP was reduced checked if turnMemory contains "NoChoice"
which caused PP to not reduce.

Note to self and others: beware of [] operator, it has side effects.