Redid the whole spriting engine
Now sprites use the same naming convention as veekun.
Subgens can easily be added:
- Replace db/pokes/generation-4.zip by database/generation-4.zip
- change pokemon-info so that the "heartgold-soulsilver/" is configurable to say, "platinum/" or "diamon-pearl/"
- enjoy!

I didn't touch gen 1 and 2 sprites because they have a white background.
To fix this, find an utility that removes the background and remove the
background of all those sprites, and then rby.zip and gsc.zip can be replaced
too.

For the future (gen 6+), in order to create an appropriate zip, download
the one from veekun and apply the ./convert-formes.sh to it and there you
go.