Removed wasted space in the teambuilder
Now one more change possible is to make gender changeable by clicking the gendericon (or make it a button with a menu attached, like the "advanced" button in the old teambuilder). If you have an opinion and believe in it, go ahead!

Now there's enough space to add hidden power selection with multiple IVS.