BattleConfiguration now complies with new network protocol
I had to keep backward compatibility with 2.0.05. You'll notice that now Analyzers have a version member, which gives the version of the client/server at the other end. That way the way messages are received/sent can be customized, depending on the version.

Todo: Add items in EngageBattle/SpectateBattle, put the rated info in the battle configuration